Add Query monitoring only flag and refactor queries file
