+++ src/queryanalysis/query_analysis.go
@@ -5,7 +5,8 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/v3/log"
 	"github.com/newrelic/nri-mssql/src/args"
 	"github.com/newrelic/nri-mssql/src/connection"
-	"github.com/newrelic/nri-mssql/src/queryanalysis/config"
+	"github.com/newrelic/nri-mssql/src/queryanalysis/queries"
+
 	"github.com/newrelic/nri-mssql/src/queryanalysis/utils"
 	"github.com/newrelic/nri-mssql/src/queryanalysis/validation"
 )
@@ -31,7 +32,7 @@ func PopulateQueryPerformanceMetrics(integration *integration.Integration, argum
 
 	utils.ValidateAndSetDefaults(&arguments)
 
-	queries := config.Queries
+	queries := queries.Queries
 	queryDetails, err := utils.LoadQueries(queries, arguments)
 	if err != nil {
 		log.Error("Error loading query configuration: %v", err)
